requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas>=0.25.3
tensorflow>=2.1.0
